examples: separate hello from capture example
